Refactor: ModalWindow.jsx 파일 위치 변경(common -> modal)
